Move referrer mapping to a dedicated class
